+++ test/session_tests.jl
@@ -26,9 +26,11 @@
     @test session.environment[:cpu_threads] == Sys.CPU_THREADS
     @test session.environment[:word_size] == Sys.WORD_SIZE
 
-    rxinfer_version = 
-        VersionNumber(TOML.parsefile(joinpath(pkgdir(RxInfer), "Project.toml"))["version"])
-    @test session.environment[:rxinfer_version] == string(rxinfer_version)
+    if get(ENV, "CI", "false") === "true"
+        # This test breaks precompilation in VSCode, thus disabled locally, executes only in CI
+        rxinfer_version = VersionNumber(TOML.parsefile(joinpath(pkgdir(RxInfer), "Project.toml"))["version"])
+        @test session.environment[:rxinfer_version] == string(rxinfer_version)
+    end
 end
 
 @testitem "RxInfer should have a default session" begin
@@ -56,220 +58,41 @@ end
     @test new_default_session.created_at > original_default_session.created_at
 end
 
-@testitem "log_data_entry" begin
-    import RxInfer: log_data_entry
-
-    @testset "Scalar values" begin
-        let entry = log_data_entry(:y => 1)
-            @test entry.name === :y
-            @test entry.type === Int
-            @test entry.size === ()
-            @test entry.elsize === ()
+@testitem "Log session should save the context" begin
+    session = RxInfer.create_session()
+    result = RxInfer.with_session(session) do invoke
+        RxInfer.append_invoke_context(invoke) do ctx
+            ctx[:a] = 1
+            ctx[:b] = 2
         end
-
-        let entry = log_data_entry(:x => 3.14)
-            @test entry.name === :x
-            @test entry.type === Float64
-            @test entry.size === ()
-            @test entry.elsize === ()
-        end
+        return 3
     end
+    @test length(session.invokes) === 1
+    last_invoke = session.invokes[end]
+    @test last_invoke.context[:a] === 1
+    @test last_invoke.context[:b] === 2
+    @test result === 3
 
-    @testset "Vectors" begin
-        let entry = log_data_entry(:x => [1])
-            @test entry.name === :x
-            @test entry.type === Vector{Int}
-            @test entry.size === (1,)
-            @test entry.elsize === ()
+    result = RxInfer.with_session(nothing) do invoke
+        RxInfer.append_invoke_context(invoke) do ctx
+            ctx[:a] = 1
+            ctx[:b] = 2
         end
-
-        let entry = log_data_entry(:x => [1.0, 2.0, 3.0])
-            @test entry.name === :x
-            @test entry.type === Vector{Float64}
-            @test entry.size === (3,)
-            @test entry.elsize === ()
-        end
-
-        let entry = log_data_entry(:x => [[1, 2], [3, 4]])
-            @test entry.name === :x
-            @test entry.type === Vector{Vector{Int}}
-            @test entry.size === (2,)
-            @test entry.elsize === (2,)
-        end
+        return 4
     end
-
-    @testset "Matrices" begin
-        let entry = log_data_entry(:x => ones(2, 3))
-            @test entry.name === :x
-            @test entry.type === Matrix{Float64}
-            @test entry.size === (2, 3)
-            @test entry.elsize === ()
-        end
-
-        let entry = log_data_entry(:x => reshape([1, 2, 3, 4], 2, 2))
-            @test entry.name === :x
-            @test entry.type === Matrix{Int}
-            @test entry.size === (2, 2)
-            @test entry.elsize === ()
-        end
-    end
-
-    @testset "Matrix of vectors" begin
-        let data = Matrix{Vector{Float64}}(undef, 2, 2)
-            data[1, 1] = [1.0, 2.0]
-            data[1, 2] = [3.0, 4.0]
-            data[2, 1] = [5.0, 6.0]
-            data[2, 2] = [7.0, 8.0]
-            entry = log_data_entry(:x => data)
-            @test entry.name === :x
-            @test entry.type === Matrix{Vector{Float64}}
-            @test entry.size === (2, 2)
-            @test entry.elsize === (2,)
-        end
-    end
-
-    struct StrangeDataEntry end
-
-    @testset let entry = log_data_entry(StrangeDataEntry)
-        @test entry.name === :unknown
-        @test entry.type === :unknown
-        @test entry.size === :unknown
-        @test entry.elsize === :unknown
-    end
+    @test result === 4
 end
 
-@testitem "log_data_entries" begin
-    import RxInfer: log_data_entry, log_data_entries
-
-    @testset "Named tuple entries" begin
-        data = (y = 1, x = [2.0, 3.0], z = [[1.0, 2.0], [3.0]])
-        entries = log_data_entries(data)
-
-        @test length(entries) === 3
-
-        # Check y entry
-        y_entry = entries[1]
-        @test y_entry.name === :y
-        @test y_entry.type === Int
-        @test y_entry.size === ()
-        @test y_entry.elsize === ()
-
-        # Check x entry
-        x_entry = entries[2]
-        @test x_entry.name === :x
-        @test x_entry.type === Vector{Float64}
-        @test x_entry.size === (2,)
-        @test x_entry.elsize === ()
-
-        # Check z entry
-        z_entry = entries[3]
-        @test z_entry.name === :z
-        @test z_entry.type === Vector{Vector{Float64}}
-        @test z_entry.size === (2,)
-        @test z_entry.elsize === (2,)
+@testitem "Log session should save errors if any" begin
+    session = RxInfer.create_session()
+    @test_throws "I'm an error" RxInfer.with_session(session) do invoke
+        error("I'm an error")
     end
+    @test length(session.invokes) === 1
+    last_invoke = session.invokes[end]
+    @test last_invoke.context[:error] == "ErrorException(\"I'm an error\")"
 
-    @testset "Dictionary entries" begin
-        data = Dict(:y => 1, :x => [2.0, 3.0], :z => [[1.0, 2.0], [3.0]])
-        entries = log_data_entries(data)
-
-        @test length(entries) === 3
-        @test Set(entry.name for entry in entries) == Set([:x, :y, :z])
-
-        # Find and check y entry
-        y_entry = findfirst(e -> e.name === :y, entries)
-        @test !isnothing(y_entry)
-        y_entry = entries[y_entry]
-        @test y_entry.type === Int
-        @test y_entry.size === ()
-        @test y_entry.elsize === ()
-
-        # Find and check x entry
-        x_entry = findfirst(e -> e.name === :x, entries)
-        @test !isnothing(x_entry)
-        x_entry = entries[x_entry]
-        @test x_entry.type === Vector{Float64}
-        @test x_entry.size === (2,)
-        @test x_entry.elsize === ()
-
-        # Find and check z entry
-        z_entry = findfirst(e -> e.name === :z, entries)
-        @test !isnothing(z_entry)
-        z_entry = entries[z_entry]
-        @test z_entry.type === Vector{Vector{Float64}}
-        @test z_entry.size === (2,)
-        @test z_entry.elsize === (2,)
+    @test_throws "I'm an error" RxInfer.with_session(nothing) do invoke
+        error("I'm an error")
     end
-
-    struct UnknownStruct end
-    @test log_data_entries(UnknownStruct()) == :unknown # be safe on something we don't know how to parse
-
-    @testset "data with UnknownStructs as elements" begin 
-        data = (y = UnknownStruct(), x = UnknownStruct())
-
-        entries = log_data_entries(data)
-
-        @test length(entries) === 2
-
-        # Check y entry
-        y_entry = entries[1]
-        @test y_entry.name === :y
-        @test y_entry.type === UnknownStruct
-        @test y_entry.size === :unknown
-        @test y_entry.elsize === :unknown
-
-        # Check x entry
-        x_entry = entries[2]
-        @test x_entry.name === :x
-        @test x_entry.type === UnknownStruct
-        @test x_entry.size === :unknown
-        @test x_entry.elsize === :unknown
-    end
-end
-
-@testitem "Session Logging basic execution" begin
-
-    # Create a simple model for testing
-    @model function simple_model(y)
-        x ~ Normal(mean = 0.0, var = 1.0)
-        y ~ Normal(mean = x, var = 1.0)
-    end
-
-    # Create test data
-    test_data = (y = 1.0,)
-
-    # Run inference inside session `session`
-    result = infer(model = simple_model(), data = test_data)
-
-    session = RxInfer.default_session()
-
-    # Basic checks, other tests may have produced more invokes here
-    @test length(session.invokes) >= 1
-
-    # Check the latest invoke
-    latest_invoke = session.invokes[end]
-    @test !isempty(latest_invoke.id)
-    @test latest_invoke.status == :success
-    @test latest_invoke.execution_end > latest_invoke.execution_start
-    @test hasproperty(latest_invoke.context, :model)
-    @test hasproperty(latest_invoke.context, :data)
-    @test !isnothing(latest_invoke.context.data)
-    @test latest_invoke.context.model == """
-    function simple_model(y)
-        x ~ Normal(0.0, 1.0)
-        y ~ Normal(x, 1.0)
-    end"""
-    @test length(latest_invoke.context.data) === 1
-
-    # Check saved properties of the passed data `y`
-    saved_data_properties = latest_invoke.context.data[end]
-    @test saved_data_properties.name === :y
-    @test saved_data_properties.type === Int
-
-    custom_session = RxInfer.create_session()
-    result = infer(model = simple_model(), data = test_data, session = session)
-
-    @test length(custom_session.invokes) === 1
-    @test latest_invoke.id != custom_session.invokes[1].id
-    @test latest_invoke.context == custom_session.invokes[1].context
 end
